Fixed Issue #5 - Dynamic loading of non-essential classes
This is handled by creating a global "Simpl" class that loads only the classes when they are requested instead of loading them all up front. This can be used when defining your own classes be sure to use "$mySimpl->Load('Email');" or what every required classes before the class definition.