Issue #25 - Added the SetOptions and SetConfig() functions to the Form() class. These can be used from anywhere in the application including the class constructor. Any function call that uses the form will overwrite the defaults on that page only.
Also fixed issue with the SetDefaults() where you could never truly ResetValues() because the defaults would always come up and then the GetList() would be skewed. To solve this all the Default references were moved only to the Form() function of the Field() class and will only be used for the Form Output.