Fixed Issues #27 & #30: Added the GetLabel() function to be able to get a label for a field outside the DbTemplate or Form class.
Added the USE_ENUM flag in the config which is defaulted to false but if set to true in the application it will search through the table and add the ENUM valse as default options for the field. It does require an extra query and should only be used with USE_CACHE turned on.