Added the SetLabels() and SetExamples() function to the Form class. These will pave the road to replace how we extend the DbTemplate to a more sane API friendly way.
Added the Pull from the DB in the DbTemplate constructor, this will eventually replace the DbTemplate() function which will eventually be deprecated. The constructor is a cleaner way to accomplish the same task. 

Set the default display to 1 in the Field constructor. If you are creating a single field outside the Form class it would not display when calling Form() since its default display was hidden.