Added a type of "settings" for forms. This seems to be an okay solution but is not the most elegant or start way of settings to manipulate core functionality. This will/should be changed to a settings file eventually but the API will continue to be similar.
Anything that inherits from Form has the "SetSetting" and "GetSetting" functions. They have just two settings to choose from "required_indicator" ('before','after') and "label_ending" (default: ':'). Both of these settings change the label output in the Form() function.

required_indicator changes where the * is for required fields, and label_ending changes what will be displayed after the label text so "First Name:" is the default, it can be changed to '' or '-' depending on need.